requirements_test.txt
python-dotenv==0.21.0
pytest==7.2.0
pytest-asyncio